Move dataset related API to the dataset module
