+++ backend/dataall/api/Objects/Group/queries.py
@@ -1,5 +1,5 @@
 from ... import gql
-from .resolvers import get_group, list_datasets_owned_by_env_group, list_data_items_shared_with_env_group, list_cognito_groups
+from .resolvers import get_group, list_data_items_shared_with_env_group, list_cognito_groups
 
 getGroup = gql.QueryField(
     name='getGroup',
@@ -8,18 +8,6 @@ getGroup = gql.QueryField(
     resolver=get_group,
 )
 
-
-listDatasetsOwnedByEnvGroup = gql.QueryField(
-    name='listDatasetsOwnedByEnvGroup',
-    type=gql.Ref('DatasetSearchResult'),
-    args=[
-        gql.Argument(name='environmentUri', type=gql.NonNullableType(gql.String)),
-        gql.Argument(name='groupUri', type=gql.NonNullableType(gql.String)),
-        gql.Argument(name='filter', type=gql.Ref('DatasetFilter')),
-    ],
-    resolver=list_datasets_owned_by_env_group,
-    test_scope='Dataset',
-)
 
 
 listDataItemsSharedWithEnvGroup = gql.QueryField(
