+++ backend/dataall/db/api/environment.py
@@ -804,75 +804,6 @@ class Environment:
         ).first()
 
     @staticmethod
-    def query_environment_datasets(session, username, groups, uri, filter) -> Query:
-        query = session.query(Dataset).filter(
-            and_(
-                Dataset.environmentUri == uri,
-                Dataset.deleted.is_(None),
-            )
-        )
-        if filter and filter.get('term'):
-            term = filter['term']
-            query = query.filter(
-                or_(
-                    Dataset.label.ilike('%' + term + '%'),
-                    Dataset.description.ilike('%' + term + '%'),
-                    Dataset.tags.contains(f'{{{term}}}'),
-                    Dataset.region.ilike('%' + term + '%'),
-                )
-            )
-        return query
-
-    @staticmethod
-    def query_environment_group_datasets(session, username, groups, envUri, groupUri, filter) -> Query:
-        query = session.query(Dataset).filter(
-            and_(
-                Dataset.environmentUri == envUri,
-                Dataset.SamlAdminGroupName == groupUri,
-                Dataset.deleted.is_(None),
-            )
-        )
-        if filter and filter.get('term'):
-            term = filter['term']
-            query = query.filter(
-                or_(
-                    Dataset.label.ilike('%' + term + '%'),
-                    Dataset.description.ilike('%' + term + '%'),
-                    Dataset.tags.contains(f'{{{term}}}'),
-                    Dataset.region.ilike('%' + term + '%'),
-                )
-            )
-        return query
-
-    @staticmethod
-    @has_resource_perm(permissions.LIST_ENVIRONMENT_DATASETS)
-    def paginated_environment_datasets(
-        session, username, groups, uri, data=None, check_perm=None
-    ) -> dict:
-        return paginate(
-            query=Environment.query_environment_datasets(
-                session, username, groups, uri, data
-            ),
-            page=data.get('page', 1),
-            page_size=data.get('pageSize', 10),
-        ).to_dict()
-
-    @staticmethod
-    def paginated_environment_group_datasets(
-        session, username, groups, envUri, groupUri, data=None, check_perm=None
-    ) -> dict:
-        return paginate(
-            query=Environment.query_environment_group_datasets(
-                session, username, groups, envUri, groupUri, data
-            ),
-            page=data.get('page', 1),
-            page_size=data.get('pageSize', 10),
-        ).to_dict()
-
-
-
-
-    @staticmethod
     def query_environment_networks(session, username, groups, uri, filter) -> Query:
         query = session.query(models.Vpc).filter(
             models.Vpc.environmentUri == uri,
