+++ backend/dataall/modules/datasets/services/dataset_service.py
@@ -623,3 +623,70 @@ class DatasetService:
             .filter(DatasetTable.datasetUri == dataset_uri)
             .count()
         )
+
+    @staticmethod
+    def query_environment_group_datasets(session, username, groups, envUri, groupUri, filter) -> Query:
+        query = session.query(Dataset).filter(
+            and_(
+                Dataset.environmentUri == envUri,
+                Dataset.SamlAdminGroupName == groupUri,
+                Dataset.deleted.is_(None),
+            )
+        )
+        if filter and filter.get('term'):
+            term = filter['term']
+            query = query.filter(
+                or_(
+                    Dataset.label.ilike('%' + term + '%'),
+                    Dataset.description.ilike('%' + term + '%'),
+                    Dataset.tags.contains(f'{{{term}}}'),
+                    Dataset.region.ilike('%' + term + '%'),
+                )
+            )
+        return query
+
+    @staticmethod
+    def query_environment_datasets(session, username, groups, uri, filter) -> Query:
+        query = session.query(Dataset).filter(
+            and_(
+                Dataset.environmentUri == uri,
+                Dataset.deleted.is_(None),
+            )
+        )
+        if filter and filter.get('term'):
+            term = filter['term']
+            query = query.filter(
+                or_(
+                    Dataset.label.ilike('%' + term + '%'),
+                    Dataset.description.ilike('%' + term + '%'),
+                    Dataset.tags.contains(f'{{{term}}}'),
+                    Dataset.region.ilike('%' + term + '%'),
+                )
+            )
+        return query
+
+    @staticmethod
+    @has_resource_perm(permissions.LIST_ENVIRONMENT_DATASETS)
+    def paginated_environment_datasets(
+            session, username, groups, uri, data=None, check_perm=None
+    ) -> dict:
+        return paginate(
+            query=DatasetService.query_environment_datasets(
+                session, username, groups, uri, data
+            ),
+            page=data.get('page', 1),
+            page_size=data.get('pageSize', 10),
+        ).to_dict()
+
+    @staticmethod
+    def paginated_environment_group_datasets(
+            session, username, groups, envUri, groupUri, data=None, check_perm=None
+    ) -> dict:
+        return paginate(
+            query=DatasetService.query_environment_group_datasets(
+                session, username, groups, envUri, groupUri, data
+            ),
+            page=data.get('page', 1),
+            page_size=data.get('pageSize', 10),
+        ).to_dict()
+
